pyproject.toml
[tool.poetry]
name = "snappy"
version = "0.1.0"
description = ""
authors = ["Connor McCarthy <mccarthy.connor.james@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/connor-mccarthy/snappy"
keywords = ["aws", "lambda", "docker", "cli"]


[tool.poetry.dependencies]
python = ">=3.7,<4.0"

cleo = "^0.8.1"
requests = "^2.25.1"
boto3 = "^1.17.27"
docker = "^4.4.4"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
pre-commit = "^2.10.1"
black = "^20.8b1"
mypy = "^0.800"
bandit = "^1.7.0"
pytest = "^6.2.2"
pytest-sugar = "^0.9.4"

[tool.poetry.scripts]
snappy = "snappy.cli.application:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.pytest]
minversion = "6.2"
addopts = "--cov=src --cov-report term-missing"
